Move and rename Config Add hashcode caching for CachedPackage Rename _dependencies
